refactor: move update util files
